+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support importing remote portable actor profiles.
+- Allow registration of portable actors with ap:// IDs.
 
 ### Changed